Enabled sync between ramping file and motor

+++ throttle/rampingCopy.cpp
@@ -1,10 +1,10 @@
 #include "rampingCopy.h"
 
 Ramping::Ramping():m_currentSpd(MIN_SPD), m_desiredSpd(MIN_SPD){}
-void Ramping::setCurrentSpd(byte spd)      {  m_currentSpd = spd; }
-void Ramping::setDesiredSpd(byte spd)      {  m_desiredSpd = spd; }
-byte Ramping::getCurrentSpd() const {  return m_currentSpd; }
-byte Ramping::getDesiredSpd() const {  return m_desiredSpd; }
+void Ramping::setCurrentSpd(int spd)      {  m_currentSpd = spd; }
+void Ramping::setDesiredSpd(int spd)      {  m_desiredSpd = spd; }
+int Ramping::getCurrentSpd() const {  return m_currentSpd; }
+int Ramping::getDesiredSpd() const {  return m_desiredSpd; }
 
 bool Ramping::isTimeToChangeSpd(long timeMillis){
   if(timeMillis - m_time >= m_maxWaitTime){
@@ -16,21 +16,21 @@ bool Ramping::isTimeToChangeSpd(long timeMillis){
   }
 }
 
-void Ramping::spdUp(byte amount){
+void Ramping::spdUp(int amount){
   if(m_currentSpd + amount >= MAX_SPD)
     m_currentSpd = MAX_SPD;
   else
     m_currentSpd += amount;
 }
 
-void Ramping::spdDown(byte amount){
+void Ramping::spdDown(int amount){
   if(m_currentSpd - amount <= MIN_SPD)
     m_currentSpd = MIN_SPD;
   else
     m_currentSpd -= amount; 
 }
 
-byte Linear::newSpd(int throttle, long timeMillis){
+int Linear::newSpd(int throttle, long timeMillis){
   setDesiredSpd(map(throttle, NEUTRAL_THROTTLE, MAX_THROTTLE, MIN_SPD, MAX_SPD));
   
   if(isTimeToChangeSpd(timeMillis)){
@@ -46,7 +46,7 @@ byte Linear::newSpd(int throttle, long timeMillis){
 
 PD::PD(): m_error(0), m_prevError(0), m_rateError(0){}
 
-byte PD::newSpd(int throttle, long timeMillis){
+int PD::newSpd(int throttle, long timeMillis){
   setDesiredSpd(map(throttle, NEUTRAL_THROTTLE, MAX_THROTTLE, MIN_SPD, MAX_SPD));
   
   if(isTimeToChangeSpd(timeMillis)){
@@ -54,7 +54,7 @@ byte PD::newSpd(int throttle, long timeMillis){
     m_rateError = m_error - m_prevError;
     m_prevError = m_error;
 
-    byte spdChange = Kp * m_error + Kd * m_rateError;
+    int spdChange = Kp * m_error + Kd * m_rateError;
   
     if(spdChange >= 0)
       spdUp(spdChange);
